Parzial Working version. Bug on solverBot.py
